Issue #948 Move EhcacheClientEntity classes to 'internal'
