+++ clustered/client/src/main/java/org/ehcache/clustered/internal/EhcacheClientEntityService.java
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehcache.clustered.client;
+
+package org.ehcache.clustered.internal;
 
 import java.util.UUID;
