+++ clustered/client/src/test/java/org/ehcache/clustered/docs/GettingStarted.java
@@ -17,7 +17,7 @@
 package org.ehcache.clustered.docs;
 
 import org.ehcache.PersistentCacheManager;
-import org.ehcache.clustered.client.UnitTestConnectionService;
+import org.ehcache.clustered.internal.UnitTestConnectionService;
 import org.ehcache.clustered.config.builders.ClusteredResourcePoolBuilder;
 import org.ehcache.clustered.config.builders.ClusteringServiceConfigurationBuilder;
 import org.ehcache.config.builders.CacheConfigurationBuilder;
