+++ clustered/client/src/test/resources/META-INF/services/org.terracotta.connection.ConnectionService
@@ -1,1 +1,1 @@
-org.ehcache.clustered.client.UnitTestConnectionService
+org.ehcache.clustered.internal.UnitTestConnectionService
