+++ clustered/integration-test/src/test/java/org/ehcache/clustered/UnitTestConnectionService.java
@@ -19,7 +19,7 @@ package org.ehcache.clustered;
 
 import java.net.URI;
 import java.util.Properties;
-import org.ehcache.clustered.client.EhcacheClientEntityService;
+import org.ehcache.clustered.internal.EhcacheClientEntityService;
 
 import org.ehcache.clustered.server.EhcacheServerEntityService;
 import org.terracotta.connection.Connection;
